시간 정지 구현 및 입력 시스템 재구성중

+++ PixelDEfect/Assets/Scripts/Weapon/WeaponBase.cs
@@ -4,12 +4,12 @@ public abstract class WeaponBase : MonoBehaviour
 {
     [Header("Weapon Settings")] 
     [SerializeField] protected string weaponName;
-    [SerializeField] protected float damage;
+    [SerializeField] protected int damage;
     [SerializeField] protected float attackCooldown;
 
     // audioSource 추가 예정
     
     public string WeaponName => weaponName;
-    public float Damage => damage;
+    public int Damage => damage;
     public float AttackCooldown => attackCooldown;
 }
